v4l: fix dv stream crash after exit playback. [1/1]
PD#SWPL-47728

Problem:
When the playback exits, the v4l2 driver does not know whether the
backend uses the relevant data in the vf, but the v4l2 driver releases
the relevant data in the dv buffer, causing the null pointer to be
released early.

Solution:
The data in the dv buffer should be released after all the vf data
is returned.

Verify:
Ohm

Change-Id: Iffbc16a39ddb78dd25bc78e4c9a30160a08792a6
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>