vdec: optimize decoder task worker. [1/1]
PD#SWPL-52287

Problem:
optimize the decode worker usage of muti-frames in a package.

Solution:
When decode processes multiple frames in a packet, will schedule
the worker infinite polling to wait buff, which will consume
a lot of CPU slice and cause to drop frames in the start playback
probability. The solution is to put the waiting task into an
independent thread to sleep and wait until the cap buff comes back.

Verify:
redi

Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>
Change-Id: Idda1ae9fdf596e39848cdece540ab02bf19343b0