media_module: CB2 v4l resolve h265 playback stuck issue [1/1]
PD#SWPL-52774

Problem:
When the user space does not send all the buffers to
the decoder, decoder will not to detect whether
the DPB is abnormal. In this case, the DPB queue is
abnormal and the playback will stuck.

Solution:
When the user space does not send all the buffers to
the decoder, it also to detect whether the DPB
queue is abnormal to solve the problem.

Verify:
SC2

Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>
Change-Id: If251fcd319ae079fc79a69080c452ec4e66e8ac1