AV1: CB1 AV1 seek freezing. [1/1]
PD#SWPL-51229

Problem:
The problem with the width and height being 0 (coded(0 x 0))
is that the sequence head may not be sent in the first frame after the seek,
but at this time the decoder still reports the width and height,
which leads to an abnormal exit.

Solution:
When the decoder does not parse the sequence head,
continue to decode until the width and height information
is parsed before reporting.

Verify:
AH212

Change-Id: I0721bce020497bc377b8032d2b788559fcec3e6b
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>