h264: CB2 Playing multiple IDR frame streams may cause jams and exit early [1/1]
PD#SWPL-53679

Problem:
There are only two I frames in the code stream and both are IDR frames.
When the second IDR frame is reached, the driver does not process the
no_output_of_prior_pics_flag according to the 264 protocol syntax.
One of the decoded buffers is not removed from the dpb,
which causes reset_buffer to be activated, but there is no I afterwards.
The frame is over, so I quit playing.

Solution:
According to the 264 protocol syntax,
no_output_of_prior_pics_flag is processed in the driver,
the reference frame flags of all frames in dpb are cancelled,
and the frames without pre_output and is_output are marked is_output.

Verify:
sc2

Signed-off-by: xinguang.jiang <xinguang.jiang@amlogic.com>
Change-Id: Ie22cb641b88353db83cb51f2119c7a46e92e935a