v4l: CB1 fixed irq-vdec-0 takes more cpu slice. [1/1]
PD#SWPL-42732

Problem:
observe frame drop after seek vp9 stream 4k 60fps.

Solution:
The root cause is that the codec mm dcache mechanism
is disabled after seek, which causes the allocation of buffer
to be very time-consuming. the solution is to ensure that
the codec mm's cache is valid.

Verify:
ah212

Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>
Change-Id: I9a7d13e8067bce1830682b4cee48e89849b6070d