vdec: CB1 wait_cap_buf not exit. [1/1]
PD#SWPL-60418

Problem:
in a multi-frame scene, the decoder is waiting for the buffer,
and the instance exits at this time,
causing the work to not finish and the hevc core is not reset.
The stream cannot be played after restarting.

Solution:
When exiting abnormally,
reset the hevc core and vdec core in vdec disconnet.

Verify:
SC2

Change-Id: Ib8d7748c1ae3be09ef737ee01081bcb32657a687
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>